requirements.txt
torch==2.1.0+cpu
transformers
accelerate
opencv-python
numpy
pyttsx3
SpeechRecognition
sounddevice
scipy
discord.py
requests
python-dotenv
